backend/app/views.py
from django.shortcuts import render, HttpResponse
from django.urls import reverse_lazy
from rest_framework.views import APIView
from rest_framework.generics import GenericAPIView
from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnly, AllowAny
from rest_framework.response import Response
from rest_framework.pagination import LimitOffsetPagination, PageNumberPagination



# Create your views here.

