backend/requirements/local.txt
#-r ./base.txt
django>=2.1.9
gunicorn==19.7.0
psycopg2==2.7.5
celery==4.3.0
redis==3.2.0
djangorestframework==3.9.1
django-cors-headers
channels==2.1.7
channels-redis==2.3.3
#daphne==2.2.5 # substitution with uvicorn
uvicorn
flower


######## ADICIONALES

##TESTING
coverage

#### DJANGO UTILS
#django-bootstrap4
#django-crispy-forms
#django-heroku
#django-extensions
#django-debug-toolbar
#django-appconf==1.0.2
#django-compressor==2.2
#django-countries==5.3.2

### GOOGLE SERVICES
#google-api-python-client==1.7.4
#google-auth==1.5.1
#google-auth-httplib2==0.0.3
#pygsheets==1.1.4
#gspread==3.0.1

### MATH AND GRAPH
#matplotlib==3.0.2
#mpl-finance==0.10.0
#numpy==1.15.0
#livedataframe
#pandas==0.23.4


### SCRAPING AND FETCH DATA
#beautifulsoup4==4.6.1
#httplib2==0.11.3
requests>=2.20.0
#selenium==3.14.0
#urllib3>=1.24.2

### PYTHON UTILS
autopep8==1.4.2
pep8==1.7.1
ipython==6.5.0
ipython-genutils==0.2.0
pylint==2.1.1
setuptools

### OTHERS
#oauth2client==4.1.2
#pyOpenSSL==18.0.0
#python-telegram-bot==10.1.0